Extracate profile menu sidebar in prep to reuse on Org pages
